Implemented diagnostics, evaluator and parser.
Diagnostics are now logged so when there is an error, such as an
unexpected token, it will be reported to the user.

Expressions can now be evaluated to produce an answer for basic math.
This takes into account operator precedence and use of parenthesis. The
parser has been expanded for this purpose.