Remove node abstract class
The Node abstract class was previously inherited by both Token and
Expression, allowing the ParseTree to iterate through an array
containing both items. This was instead changed so that the ParseTree
iterates through Objects, and simply goes down different conditionals
depending on the instance of the Object. This made the Node class
redundant.