Fix variable assignments on broken syntax
If a variable was assigned with a literal and then a non-existent variable, e.g. 'a = 1 * b', it would be assigned with the left operand by default. This has been changed so instead of returning the left operand, a null literal expression is returned. This prevents the variable from ever being assigned.