Fix assignment operators on declaration
If an assignment operator other than 'equals' was used the first time a variable is declared, it would assign it as usual without giving an error. This has been changed so that if a variable doesnt already exist in the symbol table and the assignment operator isn't 'equals', an error will be returned.