Fix unexpected token handling
Up until now the Parser was only capable of handling unexpected tokens in REPL mode, since entries are seen as a single line. In IDE mode however, entries are on individual lines and have to be checked for valid token bindings on a line-by-line basis. To avoid parsing each line individually, I introduced the concept of the 'LINE_BREAK_TOKEN'; these dont make it past the Parser and are used as an indicator to skip certain tokens when comparing bindings.

For example:
    | a = 5
    | a * 5
With the above syntax, this would have given an error because it would think '5' on line 1 and 'a' on line 2 are incompatible binds, even though they are not related expression-wise. But now the parser instead sees a 'LINE_BREAK_TOKEN' after the '5', skips ahead, and begins comparing the 'a' to the '*' instead.