Deprecated ExpressionBinder
Up until now I've been using the ExpressionBinder to define all valid token combinations. In order to simplify this and reduce the chance of bugs occurring, I've moved away from that technique and decided to instead use ValidateToken to ensure a line has ended where it should.

By passing the 'LINE_BREAK_TOKEN' to the ValidateToken method at the appropriate times, the program checks to see if there are any trailing (invalid) tokens on the same line. However, this means that problem separation between lines must be enforced for the program to run correctly, as opposed to it previously being possible to run block statements on a single line.