Rework startup
Renamed Compilation to ConsoleManager and changed its purpose so it acts as a console startup hub. Also moved it into the ui package.

Updated StageManager to be cleaner and clearer.

Update Main so that it acts as the JavaFX base class.

Made the global StageManager object in MainController local instead.